requirements.txt
torch>=2.0.0
langchain>=0.2.0
langgraph>=0.0.55
langchain-openai>=0.1.0
langchain-core>=0.2.0
nats-py>=2.6.0
dash>=2.17.0
dash-leaflet>=1.0.0
dash-bootstrap-components>=1.5.0
plotly>=5.18.0
numpy>=1.26.0
scipy>=1.13.0
pandas>=2.2.0
geopandas>=0.14.0
shapely>=2.0.0
pyproj>=3.6.0
requests>=2.31.0
aiohttp>=3.9.0
wandb>=0.16.0
pytest>=8.0.0
black>=24.0.0
flake8>=7.0.0
pyyaml>=6.0.0
python-dotenv>=1.0.0
tqdm>=4.66.0
networkx>=3.0
pgmpy>=0.1.25
ortools>=9.8
overpy>=0.7
folium>=0.16.0
pyserial>=3.5
